openai/requirements.txt
ipykernel
openai
python-dotenv
yfinance
pandas
matplotlib
numpy
colorama
streamlit
